+++ test/index.js
@@ -14,13 +14,30 @@ function eventFire (el, evtType) {
 }
 
 describe('tooltip markup', function () {
+  beforeEach(function () {
+    var tooltip = document.querySelector('.tooltip')
+
+    if (tooltip) {
+      tooltip.parentNode.removeChild(tooltip)
+    }
+  })
+
   it('should have a class of tooltip', function () {
     isotip.init()
 
     var trigger = document.querySelector('.tooltip-click')
     var tooltipTmp = isotip.open(trigger)
 
-    expect(tooltipTmp.classList.contains('tooltip')).to.equal(true)
+    expect(tooltipTmp.classList.contains('tooltip')).to.be.ok
+  })
+
+  it('should have a custom class is specified', function () {
+    isotip.init()
+
+    var trigger = document.querySelector('.tooltip-click')
+    var tooltipTmp = isotip.open(trigger)
+
+    expect(tooltipTmp.classList.contains('top-edge')).to.be.ok
   })
 
   it('should have a tooltip accent element', function () {
@@ -30,7 +47,7 @@ describe('tooltip markup', function () {
     var tooltipTmp = isotip.open(trigger)
 
     expect(tooltipTmp.childNodes[0].tagName).to.equal('DIV')
-    expect(tooltipTmp.childNodes[0].classList.contains('tooltip-accent')).to.equal(true)
+    expect(tooltipTmp.childNodes[0].classList.contains('tooltip-accent')).to.be.ok
   })
 
   it('should have a p tag containing the content', function () {
@@ -73,7 +90,7 @@ describe('tooltip markup', function () {
     var tooltipTmp = isotip.open(trigger)
 
     expect(tooltipTmp.tagName).to.equal('DIV')
-    expect(tooltipTmp.classList.contains('custom-markup')).to.equal(true)
+    expect(tooltipTmp.classList.contains('custom-markup')).to.be.ok
   })
 
   it('should use custom markup for the content if specified', function () {
@@ -83,7 +100,7 @@ describe('tooltip markup', function () {
     var tooltipTmp = isotip.open(trigger)
 
     expect(tooltipTmp.childNodes[1].tagName).to.equal('SPAN')
-    expect(tooltipTmp.childNodes[1] instanceof window.Element).to.equal(true)
+    expect(tooltipTmp.childNodes[1] instanceof window.Element).to.be.ok
   })
 
   it('should accept a DOM Element', function () {
@@ -338,7 +355,7 @@ describe('tooltip triggers', function () {
 
     eventFire(trigger, 'click')
 
-    tooltipTmp = document.querySelector('.tooltip')
+    tooltipTmp = document.querySelector('.tooltip.default')
 
     expect(tooltipTmp).to.exist
   })
@@ -352,18 +369,15 @@ describe('tooltip triggers', function () {
     })
 
     var trigger = document.querySelector('.tooltip-default')
-    var tooltipTmp
+    var tooltipTmp = isotip.open(trigger)
 
-    eventFire(trigger, 'click')
     eventFire(trigger, 'click')
 
     window.setTimeout(function () {
-      tooltipTmp = document.querySelector('.tooltip')
-
-      expect(tooltipTmp).to.not.exist
+      expect(document.documentElement.contains(tooltipTmp)).to.not.be.ok
 
       done()
-    }, isotip.options.removalDelay)
+    }, isotip.options.removalDelay + 1)
   })
 
   it('should close a tooltip on click outside of the tooltip', function (done) {
@@ -375,18 +389,15 @@ describe('tooltip triggers', function () {
     })
 
     var trigger = document.querySelector('.tooltip-default')
-    var tooltipTmp
+    var tooltipTmp = isotip.open(trigger)
 
-    eventFire(trigger, 'click')
     eventFire(document.body, 'click')
 
     window.setTimeout(function () {
-      tooltipTmp = document.querySelector('.tooltip')
-
-      expect(tooltipTmp).to.not.exist
+      expect(document.documentElement.contains(tooltipTmp)).to.not.be.ok
 
       done()
-    }, isotip.options.removalDelay)
+    }, isotip.options.removalDelay + 1)
   })
 
   it('should not close a tooltip on click on the toolip', function (done) {
@@ -398,21 +409,15 @@ describe('tooltip triggers', function () {
     })
 
     var trigger = document.querySelector('.tooltip-default')
-    var tooltipTmp
-
-    eventFire(trigger, 'click')
-
-    tooltipTmp = document.querySelector('.tooltip')
+    var tooltipTmp = isotip.open(trigger)
 
     eventFire(tooltipTmp, 'click')
 
     window.setTimeout(function () {
-      tooltipTmp = document.querySelector('.tooltip')
-
-      expect(tooltipTmp).to.exist
+      expect(document.documentElement.contains(tooltipTmp)).to.be.ok
 
       done()
-    }, isotip.options.removalDelay)
+    }, isotip.options.removalDelay + 1)
   })
 
   it('should not close a tooltip if autoClose is set to false', function (done) {
@@ -424,23 +429,15 @@ describe('tooltip triggers', function () {
     })
 
     var trigger = document.querySelector('.tooltip-click.tooltip-no-close')
-    var tooltipTmp
-
-    eventFire(trigger, 'click')
-
-    tooltipTmp = document.querySelector('.tooltip')
+    var tooltipTmp = isotip.open(trigger)
 
     eventFire(trigger, 'click')
 
     window.setTimeout(function () {
-      tooltipTmp = document.querySelector('.tooltip')
-
-      expect(tooltipTmp).to.exist
-
-      isotip.close(tooltipTmp)
+      expect(document.documentElement.contains(tooltipTmp)).to.be.ok
 
       done()
-    }, isotip.options.removalDelay)
+    }, isotip.options.removalDelay + 1)
   })
 
   it('should not close a tooltip if autoClose is set to false programmatically', function (done) {
@@ -451,24 +448,16 @@ describe('tooltip triggers', function () {
       }
     })
 
-    var trigger = document.querySelector('.tooltip-default')
+    var trigger = document.querySelector('.tooltip-click.tooltip-no-close')
     var tooltipTmp = isotip.open(trigger, { autoClose: false })
 
     eventFire(trigger, 'click')
 
-    tooltipTmp = document.querySelector('.tooltip')
-
-    eventFire(trigger, 'click')
-
     window.setTimeout(function () {
-      tooltipTmp = document.querySelector('.tooltip')
-
-      expect(tooltipTmp).to.exist
-
-      isotip.close(tooltipTmp)
+      expect(document.documentElement.contains(tooltipTmp)).to.be.ok
 
       done()
-    }, isotip.options.removalDelay)
+    }, isotip.options.removalDelay + 1)
   })
 
   it('should not close a tooltip on click on a child element in toolip', function (done) {
@@ -479,24 +468,17 @@ describe('tooltip triggers', function () {
       }
     })
 
-    var trigger = document.querySelector('.tooltip-default')
-    var tooltipTmp
-    var tooltipContent
-
-    eventFire(trigger, 'click')
-
-    tooltipTmp = document.querySelector('.tooltip')
-    tooltipContent = tooltipTmp.querySelector('.tooltip-content')
+    var trigger = document.querySelector('.tooltip-html')
+    var tooltipTmp = isotip.open(trigger)
+    var tooltipContent = tooltipTmp.querySelector('.custom')
 
     eventFire(tooltipContent, 'click')
 
     window.setTimeout(function () {
-      tooltipTmp = document.querySelector('.tooltip')
-
-      expect(tooltipTmp).to.exist
+      expect(document.documentElement.contains(tooltipTmp)).to.be.ok
 
       done()
-    }, isotip.options.removalDelay)
+    }, isotip.options.removalDelay + 1)
   })
 
   it('should only remove the tooltip after the removal delay time', function (done) {
@@ -514,17 +496,15 @@ describe('tooltip triggers', function () {
 
     eventFire(trigger, 'click')
 
-    tooltipTmp = document.querySelector('.tooltip')
+    tooltipTmp = document.querySelector('.tooltip.default')
 
-    expect(tooltipTmp).to.exist
+    expect(document.documentElement.contains(tooltipTmp)).to.be.ok
 
     window.setTimeout(function () {
-      tooltipTmp = document.querySelector('.tooltip')
-
-      expect(tooltipTmp).to.not.exist
+      expect(document.documentElement.contains(tooltipTmp)).to.not.be.ok
 
       done()
-    }, isotip.options.removalDelay)
+    }, isotip.options.removalDelay + 1)
   })
 
   it('should open a tooltip on hover', function () {
@@ -560,12 +540,12 @@ describe('tooltip triggers', function () {
     eventFire(trigger, 'mouseout')
 
     window.setTimeout(function () {
-      tooltipTmp = document.querySelector('.tooltip')
+      tooltipTmp = document.querySelector('.tooltip.hover')
 
       expect(tooltipTmp).to.not.exist
 
       done()
-    }, isotip.options.removalDelay)
+    }, isotip.options.removalDelay + 1)
   })
 
   it('should not close a tooltip on mouseout if hovered over the tooltip', function (done) {
@@ -581,15 +561,15 @@ describe('tooltip triggers', function () {
 
     eventFire(trigger, 'mouseover')
 
-    tooltipTmp = document.querySelector('.tooltip')
+    tooltipTmp = document.querySelector('.tooltip.hover')
 
     eventFire(tooltipTmp, 'mouseover')
 
     window.setTimeout(function () {
-      expect(tooltipTmp.parentNode.nodeName).to.equal('BODY')
+      expect(document.documentElement.contains(tooltipTmp)).to.be.ok
 
       done()
-    }, isotip.options.removalDelay)
+    }, isotip.options.removalDelay + 1)
   })
 
   it('should not close a tooltip on mouseout if autoClose is set to false', function (done) {
@@ -601,20 +581,15 @@ describe('tooltip triggers', function () {
     })
 
     var trigger = document.querySelector('.tooltip-hover.tooltip-no-close')
-    var tooltipTmp
+    var tooltipTmp = isotip.open(trigger)
 
-    eventFire(trigger, 'mouseover')
     eventFire(trigger, 'mouseout')
 
     window.setTimeout(function () {
-      tooltipTmp = document.querySelector('.tooltip')
-
-      expect(tooltipTmp).to.exist
-
-      isotip.close(tooltipTmp)
+      expect(document.documentElement.contains(tooltipTmp)).to.be.ok
 
       done()
-    }, isotip.options.removalDelay)
+    }, isotip.options.removalDelay + 1)
   })
 
   it('should not close a tooltip on mouseout if autoClose is set to false programmatically', function (done) {
@@ -625,20 +600,16 @@ describe('tooltip triggers', function () {
       }
     })
 
-    var trigger = document.querySelector('.tooltip-hover')
+    var trigger = document.querySelector('.tooltip-hover.tooltip-no-close')
     var tooltipTmp = isotip.open(trigger, { autoClose: false })
 
     eventFire(trigger, 'mouseout')
 
     window.setTimeout(function () {
-      tooltipTmp = document.querySelector('.tooltip')
-
-      expect(tooltipTmp).to.exist
-
-      isotip.close(tooltipTmp)
+      expect(document.documentElement.contains(tooltipTmp)).to.be.ok
 
       done()
-    }, isotip.options.removalDelay)
+    }, isotip.options.removalDelay + 1)
   })
 
   it('should open a tooltip on focus', function () {
@@ -654,7 +625,7 @@ describe('tooltip triggers', function () {
 
     eventFire(trigger, 'focus')
 
-    tooltipTmp = document.querySelector('.tooltip')
+    tooltipTmp = document.querySelector('.tooltip.focus')
 
     expect(tooltipTmp).to.exist
   })
@@ -674,12 +645,12 @@ describe('tooltip triggers', function () {
     eventFire(trigger, 'blur')
 
     window.setTimeout(function () {
-      tooltipTmp = document.querySelector('.tooltip')
+      tooltipTmp = document.querySelector('.tooltip.focus')
 
-      expect(tooltipTmp).to.not.exist
+      expect(document.documentElement.contains(tooltipTmp)).to.not.be.ok
 
       done()
-    }, isotip.options.removalDelay)
+    }, isotip.options.removalDelay + 1)
   })
 
   it('should not close a tooltip on blur if autoClose is set to false', function (done) {
@@ -691,20 +662,15 @@ describe('tooltip triggers', function () {
     })
 
     var trigger = document.querySelector('.tooltip-focus.tooltip-no-close')
-    var tooltipTmp
+    var tooltipTmp = isotip.open(trigger)
 
-    eventFire(trigger, 'focus')
     eventFire(trigger, 'blur')
 
     window.setTimeout(function () {
-      tooltipTmp = document.querySelector('.tooltip')
-
-      expect(tooltipTmp).to.exist
-
-      isotip.close(tooltipTmp)
+      expect(document.documentElement.contains(tooltipTmp)).to.be.ok
 
       done()
-    }, isotip.options.removalDelay)
+    }, isotip.options.removalDelay + 1)
   })
 
   it('should not close a tooltip on blur if autoClose is set to false programmatically', function (done) {
@@ -721,13 +687,9 @@ describe('tooltip triggers', function () {
     eventFire(trigger, 'blur')
 
     window.setTimeout(function () {
-      tooltipTmp = document.querySelector('.tooltip')
-
-      expect(tooltipTmp).to.exist
-
-      isotip.close(tooltipTmp)
+      expect(document.documentElement.contains(tooltipTmp)).to.be.ok
 
       done()
-    }, isotip.options.removalDelay)
+    }, isotip.options.removalDelay + 1)
   })
 })